+++ releasenotes.md
@@ -8,6 +8,7 @@
 - Inline associated enum values are deprecated, use `--use-old-enums` to re-enable them.
 - `$typeof` may return a compile time type.
 - Improved error messages on missing qualifier on enum value. #2260
+- Add `--echo-prefix` to edit the prefix with `$echo` statements. Supports {FILE} and {LINE}
 
 ### Fixes
 - mkdir/rmdir would not work properly with substring paths on non-windows platforms.